Fixed up tournament system, not fully finished yet.

+++ src/main/java/com/rs/kotlin/game/world/activity/pvpgame/PvPGame.kt
@@ -5,14 +5,17 @@ import com.rs.java.game.map.MapBuilder
 import com.rs.java.game.player.Player
 import com.rs.java.game.player.actions.combat.Magic
 import com.rs.java.utils.Utils
+import com.rs.kotlin.game.world.activity.pvpgame.tournament.TournamentGame
 
 abstract class PvPGame {
     protected var boundChunks: IntArray? = null
     protected val players = mutableListOf<Player>()
-    protected lateinit var area: PvPAreaType
+    protected var area: PvPAreaType? = null   // <-- was lateinit
 
     abstract fun start()
     abstract fun onPlayerDeath(player: Player)
+
+    protected open fun shouldDestroyArea(): Boolean = true
 
     protected fun allocateArea(area: PvPAreaType) {
         this.area = area
@@ -32,20 +35,28 @@ abstract class PvPGame {
     fun cleanup(winner: Player?) {
         players.forEach {
             it.closePvPOverlay()
-            if (winner != null) {
-                it.showResult(winner)
-            }
-            Magic.teleport(it, WorldTile(2992, 9676, 0)) // Clan wars lobby
+
+            if (winner != null) it.showResult(winner)
+
+            // default behavior for most games
+            Magic.teleport(it, WorldTile(2992, 9676, 0))
+            it.activePvPGame = null
         }
-        players.forEach { it.activePvPGame = null }
-        destroyArea()
+
+        if (shouldDestroyArea()) {
+            destroyArea()
+        }
+
         PvPGameManager.unregisterGame(this)
     }
 
     protected fun destroyArea() {
-        val width = (area.northEastTile.x - area.southWestTile.x) / 8 + 1
-        val height = (area.northEastTile.y - area.southWestTile.y) / 8 + 1
-        MapBuilder.destroyMap(boundChunks!![0], boundChunks!![1], width, height)
+        val a = area ?: return
+        val chunks = boundChunks ?: return
+
+        val width = (a.northEastTile.x - a.southWestTile.x) / 8 + 1
+        val height = (a.northEastTile.y - a.southWestTile.y) / 8 + 1
+        MapBuilder.destroyMap(chunks[0], chunks[1], width, height)
     }
 
     protected fun randomSpawn(): WorldTile {
